pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "numpy>=2.3.3",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.11.0",
    "pytest-asyncio>=1.2.0",
    "pytest>=8.4.2",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.1.1",
    "rapidfuzz>=3.14.1",
    "scikit-learn>=1.7.2",
    "tenacity>=9.1.2",
    "uvicorn>=0.37.0",
]
